requirements.txt
streamlit>=1.32.0
google-cloud-firestore>=2.14.0
google-cloud-aiplatform>=1.49.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.20.0
folium>=0.16.0
streamlit-folium>=0.20.0
